// src/app/User/types/UserType.ts
export interface IUser {
  firstName: string;
  lastName: string;
  password: string;
  photo: string;
  totalRating: number;
  rating: number;
  categories: string[];
  active: boolean;
  phone: string;
  rule: string;
  isWorker: boolean;
  address: string;
  birthOfDate: any;
  location: any;
  subscriptionDate: any;
  experience: string;
}